Added to Working Surface Style
Provided examples of TC Needle Holders, one with a Smooth Working Surface and one with a Cross-Serrated Working Surface.  Provided links.
Provided examples of TC Forceps with a Cross-Serrated Working Surface.  Provided link.
Not able to locate an example of TC Forceps with a smooth working surface at this time.